### 2024-10-24 #7409 @editorialbot
[PRE REVIEW]: Vector: JIT-compilable mathematical manipulations of ragged Lorentz vectors
**Submitting author:** @Saransh-cpp (<a href="http://orcid.org/0000-0003-3046-7675">Saransh Chopra</a>)
**Repository:** https://github.com/scikit-hep/vector
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.5.2
