### 2024-10-27 bfb104b
Remove comments in examples and update output statements
