### 2024-10-27 2551843
Added examples for azimuthal and longitudinal properties in 3D and 4D (#519)
* Added examples for azimuthal and longitudinal properties in 3D and 4D

* Remove comments in examples and update output statements

* Apply suggestions from code review

---------

Co-authored-by: Saransh Chopra <saransh0701@gmail.com>
