### 2024-10-28 #7409 @matthewfeickert
> Our go-to HEP editor @matthewfeickert has a COI here, unfortunately.

To any would-be-editors considering this submission, my (highly biased) view is that it will be a super easy review to direct. :D
