### 2024-11-01 #7409 @kyleniemeyer
Thanks! (the `editorialbot` command has to be the first line in a comment)
