+++ .pre-commit-config.yaml
@@ -19,7 +19,7 @@ repos:
       - id: trailing-whitespace
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.7.2"
+    rev: "v0.7.3"
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
