+++ .github/workflows/ci.yml
@@ -93,7 +93,7 @@ jobs:
         run: nox -s coverage-${{ matrix.python-version }} --verbose
 
       - name: Upload coverage report
-        uses: codecov/codecov-action@v4.6.0
+        uses: codecov/codecov-action@v5.0.2
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
 
