+++ .github/workflows/ci.yml
@@ -28,7 +28,7 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: 3.x
-      - uses: astral-sh/setup-uv@v3
+      - uses: astral-sh/setup-uv@v4
       - name: PyLint
         run: uvx nox -s pylint -- --output-format=github
 
@@ -53,7 +53,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
           allow-prereleases: true
 
-      - uses: astral-sh/setup-uv@v3
+      - uses: astral-sh/setup-uv@v4
 
       - name: Test lite package
         run: uvx nox -s lite-${{ matrix.python-version }} --verbose
@@ -80,7 +80,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
           allow-prereleases: true
 
-      - uses: astral-sh/setup-uv@v3
+      - uses: astral-sh/setup-uv@v4
 
       - name: Install nox
         run: uv tool install nox
@@ -93,7 +93,7 @@ jobs:
         run: nox -s coverage-${{ matrix.python-version }} --verbose
 
       - name: Upload coverage report
-        uses: codecov/codecov-action@v5.0.2
+        uses: codecov/codecov-action@v5.0.7
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
 
@@ -107,7 +107,7 @@ jobs:
         with:
           python-version: 3.8.13
 
-      - uses: astral-sh/setup-uv@v3
+      - uses: astral-sh/setup-uv@v4
 
       - name: Check compute features
         run: uv pip install --system .[test,test-extras]
