+++ src/vector/backends/object.py
@@ -727,9 +727,8 @@ class VectorObject2D(VectorObject, Planar, Vector2D):
             azcoords = _coord_object_type[returns[0]](result[0], result[1])
             return cls.ProjectionClass2D(azimuthal=azcoords)
 
-        elif (
-            len(returns) == 2
-            or (len(returns) == 3 and returns[2] is None)
+        elif len(returns) == 2 or (
+            (len(returns) == 3 and returns[2] is None)
             and isinstance(returns[0], type)
             and issubclass(returns[0], Azimuthal)
             and isinstance(returns[1], type)
@@ -1129,9 +1128,8 @@ class VectorObject3D(VectorObject, Spatial, Vector3D):
             azcoords = _coord_object_type[returns[0]](result[0], result[1])
             return cls.ProjectionClass2D(azimuthal=azcoords)
 
-        elif (
-            len(returns) == 2
-            or (len(returns) == 3 and returns[2] is None)
+        elif len(returns) == 2 or (
+            (len(returns) == 3 and returns[2] is None)
             and isinstance(returns[0], type)
             and issubclass(returns[0], Azimuthal)
             and isinstance(returns[1], type)
