### 2024-12-02 77f2509
chore: update pre-commit hooks
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.8.0 → v0.8.1](https://github.com/astral-sh/ruff-pre-commit/compare/v0.8.0...v0.8.1)
- [github.com/rbubley/mirrors-prettier: v3.3.3 → v3.4.1](https://github.com/rbubley/mirrors-prettier/compare/v3.3.3...v3.4.1)
