### 2024-12-03 a440184
Adds a nox session for disassemble tests
Fixes #532
