### 2024-12-03 356ece5
removes pytest options set by package default and adds verbosity
Co-authored-by: Saransh Chopra <saransh0701@gmail.com>
