+++ .github/workflows/ci.yml
@@ -110,7 +110,7 @@ jobs:
       - uses: astral-sh/setup-uv@v4
 
       - name: Test compute features
-        run: uvx nox -s disassemble
+        run: uvx nox -s disassemble --verbose
 
   pass:
     needs: [pre-commit, check-lite, check-full, discheck]
