### 2024-12-03 599dd0d
Adds a nox session for disassemble tests (#534)
* Adds a nox session for disassemble tests

Fixes #532

* removes pytest options set by package default and adds verbosity

Co-authored-by: Saransh Chopra <saransh0701@gmail.com>

---------

Co-authored-by: Saransh Chopra <saransh0701@gmail.com>
