### 2024-12-07 cddd72b
Update docs/index.rst
