### 2024-12-07 b3b909e
Fix : Issue in readMe and docs/index.rst file link not working (#537)
* fix: updated links in readme.me

* fix : typo

* fix : add broken link

* Update docs/index.rst

---------

Co-authored-by: Saransh Chopra <saransh0701@gmail.com>
