### 2024-12-09 fc63268
chore: update pre-commit hooks
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.8.1 → v0.8.2](https://github.com/astral-sh/ruff-pre-commit/compare/v0.8.1...v0.8.2)
- [github.com/rbubley/mirrors-prettier: v3.4.1 → v3.4.2](https://github.com/rbubley/mirrors-prettier/compare/v3.4.1...v3.4.2)
