### 2024-12-12 e897a03
force touch columns needed to build internal vector classes
