+++ tests/backends/test_dask_awkward.py
@@ -23,3 +23,17 @@ def test_constructor():
     assert isinstance(vec.compute(), vector.backends.awkward.VectorAwkward2D)
     assert ak.all(vec.x.compute() == ak.Array([1, 1.1]))
     assert ak.all(vec.y.compute() == ak.Array([2, 2.2]))
+
+
+def test_necessary_columns():
+    vec = vector.Array([[{"pt": 1, "phi": 2}], [], [{"pt": 3, "phi": 4}]])
+    dak_vec = dak.from_awkward(vec, npartitions=1)
+
+    cols = next(iter(dak.report_necessary_columns(dak_vec).values()))
+
+    # this may seem weird at first: why would you need "phi" and "rho", if you asked for "pt"?
+    # the reason is that vector will build internally a class with "phi" and "rho",
+    # see: https://github.com/scikit-hep/vector/blob/main/src/vector/backends/awkward.py#L166-L167
+    # So, even if you ask for "pt", you will need "phi" and "rho" as well in order to build the vector class in the first place.
+    # (the same argument holds true for all other vector classes)
+    assert cols == frozenset({"phi", "rho"})
