### 2024-12-12 e41c7ed
touch only typetracer arrays, improve comment for test
