### 2024-12-18 8479c05
remove the old; I'll use vector.readthedocs.io as a guide for refilling it
