+++ docs/api/compute/planar/vector._compute.planar.deltaphi.rst
@@ -1,8 +1,0 @@
-vector._compute.planar.deltaphi module
-======================================
-
-.. automodule:: vector._compute.planar.deltaphi
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
