+++ docs/api/compute/spatial/vector._compute.spatial.z.rst
@@ -1,8 +1,0 @@
-vector._compute.spatial.z module
-================================
-
-.. automodule:: vector._compute.spatial.z
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
