+++ docs/api/compute/lorentz/vector._compute.lorentz.to_beta3.rst
@@ -1,8 +1,0 @@
-vector._compute.lorentz.to_beta3 module
-=======================================
-
-.. automodule:: vector._compute.lorentz.to_beta3
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
