+++ docs/api/compute/planar/vector._compute.planar.is_parallel.rst
@@ -1,8 +1,0 @@
-vector._compute.planar.is_parallel module
-=========================================
-
-.. automodule:: vector._compute.planar.is_parallel
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
