+++ docs/api/compute/lorentz/vector._compute.lorentz.is_timelike.rst
@@ -1,8 +1,0 @@
-vector._compute.lorentz.is_timelike module
-==========================================
-
-.. automodule:: vector._compute.lorentz.is_timelike
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
