+++ docs/api/compute/lorentz/vector._compute.lorentz.Mt2.rst
@@ -1,8 +1,0 @@
-vector._compute.lorentz.Mt2 module
-==================================
-
-.. automodule:: vector._compute.lorentz.Mt2
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
