+++ docs/api/compute/lorentz/vector._compute.lorentz.scale.rst
@@ -1,8 +1,0 @@
-vector._compute.lorentz.scale module
-====================================
-
-.. automodule:: vector._compute.lorentz.scale
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
