+++ docs/api/compute/spatial/vector._compute.spatial.mag.rst
@@ -1,8 +1,0 @@
-vector._compute.spatial.mag module
-==================================
-
-.. automodule:: vector._compute.spatial.mag
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
