+++ docs/api/compute/lorentz/vector._compute.lorentz.transform4D.rst
@@ -1,8 +1,0 @@
-vector._compute.lorentz.transform4D module
-==========================================
-
-.. automodule:: vector._compute.lorentz.transform4D
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
