+++ docs/api/compute/planar/vector._compute.planar.add.rst
@@ -1,8 +1,0 @@
-vector._compute.planar.add module
-=================================
-
-.. automodule:: vector._compute.planar.add
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :private-members:
