+++ docs/index.rst
@@ -12,53 +12,16 @@
 Overview
 --------
 
-Vector is a Python 3.8+ library (Python 3.6 and 3.7 supported till ``v0.9.0`` and ``v1.0.0``, respectively) for 2D, 3D, and `Lorentz vectors <https://en.wikipedia.org/wiki/Special_relativity#Physics_in_spacetime>`_, especially *arrays of vectors*, to solve common physics problems in a NumPy-like way.
-
-Main features of Vector:
-
-* Pure Python with NumPy as its only dependency. This makes it easier to install.
-* Vectors may be represented in a variety of coordinate systems: Cartesian, cylindrical, pseudorapidity, and any combination of these with time or proper time for Lorentz vectors. In all, there are 12 coordinate systems: {*x* - *y* vs *ρ* - *φ* in the azimuthal plane} × {*z* vs *θ* vs *η* longitudinally} × {*t* vs *τ* temporally}.
-* Uses names and conventions set by `ROOT <https://root.cern/>`_'s `TLorentzVector <https://root.cern.ch/doc/master/classTLorentzVector.html>`_ and `Math::LorentzVector <https://root.cern.ch/doc/master/classROOT_1_1Math_1_1LorentzVector.html>`_, as well as `scikit-hep/math <https://github.com/scikit-hep/scikit-hep/tree/master/skhep/math>`_, `uproot-methods TLorentzVector <https://github.com/scikit-hep/uproot3-methods/blob/master/uproot3_methods/classes/TLorentzVector.py>`_, `henryiii/hepvector <https://github.com/henryiii/hepvector>`_, and `coffea.nanoevents.methods.vector <https://coffea-hep.readthedocs.io/en/latest/modules/coffea.nanoevents.methods.vector.html>`_ (backed by Scikit-HEP/vector since ``v2024.8``).
-* Implemented on a variety of backends:
-   * pure Python objects
-   * `SymPy <https://www.sympy.org/en/index.html>`_ vectors
-   * NumPy arrays of vectors (as a `structured array <https://numpy.org/doc/stable/user/basics.rec.html>`_ subclass)
-   * `Awkward Arrays <https://awkward-array.org/>`_ of vectors
-   * potential for more: CuPy, TensorFlow, Torch...
-* Awkward and Object backends also implemented in `Numba <https://numba.pydata.org/>`_ for JIT-compiled calculations on vectors.
-* `JAX <https://awkward-array.org/doc/main/user-guide/how-to-specialize-differentiate-jax.html>`_ and `Dask <https://dask-awkward.readthedocs.io/en/stable/>`_ support through Awkward Arrays.
-* Distinction between geometrical vectors, which have a minimum of attribute and method names, and vectors representing momentum, which have synonyms like ``pt`` = ``rho``, ``energy`` = ``t``, ``mass`` = ``tau``.
-
 Installation
 ------------
-Vector is available on `PyPI <https://pypi.org/project/vector/>`_ as well as on `conda <https://anaconda.org/conda-forge/vector>`_. The library can be installed using ``pip`` -
 
-.. code-block::
+Getting help
+------------
 
-   pip install vector
+Contributing to Vector
+----------------------
 
-or using ``conda`` -
-
-.. code-block::
-
-   conda install -c conda-forge vector
-
-Example gallery
----------------
-``Vector`` has several examples covering the basics as well as some advanced usage of the library. The example gallery
-covers almost all the features offered by ``vector`` and any new additions to the gallery are welcomed.
-
-**Note**: Adding more examples and improving the existing examples for newcomers is still in progress.
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Example gallery
-
-   usage/intro
-   usage/structure
-   usage/vector_design_prototype
-
-Talks about vector
+Talks about Vector
[…]
