+++ docs/index.md
@@ -1,0 +1,51 @@
+![](_images/vector-logo.png)
+
+[![Actions Status][actions-badge]][actions-link]
+[![Documentation Status][rtd-badge]][rtd-link]
+[![pre-commit.ci status][pre-commit-badge]][pre-commit-link]
+[![codecov percentage][codecov-badge]][codecov-link]
+[![GitHub Discussion][github-discussions-badge]][github-discussions-link]
+[![Gitter][gitter-badge]][gitter-link]
+
+[![PyPI platforms][pypi-platforms]][pypi-link]
+[![PyPI version][pypi-version]][pypi-link]
+[![Conda latest release][conda-version]][conda-link]
+[![DOI][zenodo-badge]][zenodo-link]
+[![LICENSE][license-badge]][license-link]
+[![Scikit-HEP][sk-badge]][sk-link]
+
+# Overview
+
+# Installation
+
+# Getting help
+
+# Contributing to Vector
+
+```{toctree}
+src/talks.md
+```
+
+[actions-badge]: https://github.com/scikit-hep/vector/actions/workflows/ci.yml/badge.svg
+[actions-link]: https://github.com/scikit-hep/vector/actions
+[codecov-badge]: https://codecov.io/gh/scikit-hep/vector/branch/main/graph/badge.svg?token=YBv60ueORQ
+[codecov-link]: https://codecov.io/gh/scikit-hep/vector
+[conda-version]: https://img.shields.io/conda/vn/conda-forge/vector.svg
+[conda-link]: https://github.com/conda-forge/vector-feedstock
+[github-discussions-badge]: https://img.shields.io/static/v1?label=Discussions&message=Ask&color=blue&logo=github
+[github-discussions-link]: https://github.com/scikit-hep/vector/discussions
+[gitter-badge]: https://badges.gitter.im/Scikit-HEP/vector.svg
+[gitter-link]: https://gitter.im/Scikit-HEP/vector?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge
+[license-badge]: https://img.shields.io/badge/License-BSD_3--Clause-blue.svg
+[license-link]: https://opensource.org/licenses/BSD-3-Clause
+[pre-commit-badge]: https://results.pre-commit.ci/badge/github/scikit-hep/vector/main.svg
+[pre-commit-link]: https://results.pre-commit.ci/repo/github/scikit-hep/vector
+[pypi-link]: https://pypi.org/project/vector/
+[pypi-platforms]: https://img.shields.io/pypi/pyversions/vector
+[pypi-version]: https://badge.fury.io/py/vector.svg
+[rtd-badge]: https://readthedocs.org/projects/vector/badge/?version=latest
+[rtd-link]: https://vector.readthedocs.io/en/latest/?badge=latest
+[sk-badge]: https://scikit-hep.org/assets/images/Scikit--HEP-Project-blue.svg
+[sk-link]: https://scikit-hep.org/
+[zenodo-badge]: https://zenodo.org/badge/DOI/10.5281/zenodo.7054478.svg
+[zenodo-link]: https://doi.org/10.5281/zenodo.7054478

[…]
