### 2024-12-18 2fc82ac
skip test of non-existent notebook
