### 2024-12-18 bab2720
simplified the Overview page
