### 2024-12-18 1765082
style: pre-commit fixes
