+++ docs/index.md
@@ -30,9 +30,9 @@ Vector is a Python library for 2D and 3D spatial vectors, as well as 4D space-ti
 
 Vectors may be expressed in any of these coordinate systems:
 
-* the azimuthal plane may be Cartesian `x` `y` or polar `rho` ($\rho$) `phi` ($\phi$)
-* the longitudinal axis may be Cartesian `z`, polar `theta` ($\theta$), or pseudorapidity `eta` ($\eta$)
-* the temporal component for space-time vectors may be Cartesian `t` or proper time `tau` ($\tau$)
+- the azimuthal plane may be Cartesian `x` `y` or polar `rho` ($\rho$) `phi` ($\phi$)
+- the longitudinal axis may be Cartesian `z`, polar `theta` ($\theta$), or pseudorapidity `eta` ($\eta$)
+- the temporal component for space-time vectors may be Cartesian `t` or proper time `tau` ($\tau$)
 
 in any combination. (That is, 4D vectors have 2×3×2 = 12 distinct coordinate systems.)
 
@@ -40,19 +40,19 @@ in any combination. (That is, 4D vectors have 2×3×2 = 12 distinct coordinate s
 
 Vectors may be included in any of these data types:
 
-* `vector.obj` objects (pure Python)
-* [NumPy structured arrays](https://numpy.org/doc/stable/user/basics.rec.html) of vectors
-* [Awkward Arrays](https://awkward-array.org/) of vectors (possibly within variable-length lists or nested record structures)
-* [SymPy expressions](https://www.sympy.org/en/index.html) for symbolic (non-numeric) manipulations
+- `vector.obj` objects (pure Python)
+- [NumPy structured arrays](https://numpy.org/doc/stable/user/basics.rec.html) of vectors
+- [Awkward Arrays](https://awkward-array.org/) of vectors (possibly within variable-length lists or nested record structures)
+- [SymPy expressions](https://www.sympy.org/en/index.html) for symbolic (non-numeric) manipulations
 
 Each of these "backends" includes the same suite of functions.
 
 ## Getting help
 
-* Source code on GitHub: [scikit-hep/vector](https://github.com/scikit-hep/vector)
-* Report bugs and request features on the [GitHub Issues page](https://github.com/scikit-hep/vector/issues)
-* Ask questions on the [GitHub Discussions page](https://github.com/scikit-hep/vector/discussions)
-* Real-time chat on Gitter: [Scikit-HEP/Vector](https://gitter.im/Scikit-HEP/vector)
+- Source code on GitHub: [scikit-hep/vector](https://github.com/scikit-hep/vector)
+- Report bugs and request features on the [GitHub Issues page](https://github.com/scikit-hep/vector/issues)
+- Ask questions on the [GitHub Discussions page](https://github.com/scikit-hep/vector/discussions)
+- Real-time chat on Gitter: [Scikit-HEP/Vector](https://gitter.im/Scikit-HEP/vector)
 
 ## Contributing to Vector
 
