### 2024-12-18 ccb55e1
also mention geometric vs momentum
