### 2024-12-18 2f84f41
added interfaces
