+++ docs/src/common.md
@@ -1,0 +1,13 @@
+# Interface for all vectors
+
+Vectors of all backends, 2D/3D/4D, geometric and momentum, have the following attributes, properties, and methods.
+
+For interfaces specialized to 2D/3D/4D or momentum vectors, see the sections after this one.
+
+```{eval-rst}
+.. autoclass:: vector._methods.VectorProtocol
+    :members:
+    :inherited-members:
+    :member-order: bysource
+    :exclude-members: lib,ProjectionClass2D,ProjectionClass3D,ProjectionClass4D,GenericClass,MomentumClass
+```
