### 2024-12-19 fd1c652
style: pre-commit fixes
