### 2024-12-19 6994385
made stubs for tutorials
