### 2024-12-19 294db57
use SVG and make README match readthedocs
