+++ README.md
@@ -1,4 +1,4 @@
-<img alt="Vector logo" width="433" src="https://raw.githubusercontent.com/scikit-hep/vector/main/docs/_images/vector-logo.png"/>
+<img alt="Vector logo" width="50%" src="https://raw.githubusercontent.com/scikit-hep/vector/main/docs/_images/LogoSrc.svg"/>
 
 # Vector: arrays of 2D, 3D, and Lorentz vectors
 
@@ -16,13 +16,106 @@
 [![LICENSE][license-badge]][license-link]
 [![Scikit-HEP][sk-badge]][sk-link]
 
-## Overview
-
 ## Installation
+
+You can install Vector with [pip](https://pypi.org/project/vector/) and [conda](https://anaconda.org/conda-forge/vector).
+
+```bash
+pip install vector
+```
+
+## Introduction
+
+Vector is a Python library for 2D and 3D spatial vectors, as well as 4D space-time vectors. It is especially intended for performing geometric calculations on _arrays of vectors_, rather than one vector at a time in a Python for loop.
+
+Vector is part of the [Scikit-HEP project](https://scikit-hep.org/), High Energy Physics (HEP) tools in Python.
+
+### Coordinate systems
+
+Vectors may be expressed in any of these coordinate systems:
+
+- the azimuthal plane may be Cartesian `x` `y` or polar `rho` ($\rho$) `phi` ($\phi$)
+- the longitudinal axis may be Cartesian `z`, polar `theta` ($\theta$), or pseudorapidity `eta` ($\eta$)
+- the temporal component for space-time vectors may be Cartesian `t` or proper time `tau` ($\tau$)
+
+in any combination. (That is, 4D vectors have 2×3×2 = 12 distinct coordinate systems.)
+
+<img alt="Diagram of coordinate systems" width="100%" src="https://raw.githubusercontent.com/scikit-hep/vector/main/docs/_images/coordinate-systems.svg"/>
+
+### Backends
+
+Vectors may be included in any of these data types:
+
+- [vector.obj](https://vector.readthedocs.io/en/latest/src/make_object.html) objects (pure Python)
+- [NumPy structured arrays](https://numpy.org/doc/stable/user/basics.rec.html) of vectors
+- [Awkward Arrays](https://awkward-array.org/) of vectors (possibly within variable-length lists or nested record structures)
+- [SymPy expressions](https://www.sympy.org/en/index.html) for symbolic (non-numeric) manipulations
+- In [Numba-compiled functions](https://numba.pydata.org/), with [vector.obj](https://vector.readthedocs.io/en/latest/src/make_object.html) objects or Awkward Arrays
+
+Each of these "backends" provides the same suite of properties and methods, through a common "compute" library.
+
+### Geometric versus momentum
+
+Finally, vectors come in two flavors:
+
+- geometric: only one name for each property or method
+- momentum: same property or method can be accessed with several synonyms, such as `pt` ($p_T$, transverse momentum) for the azimuthal magnitude `rho` ($\rho$) and `energy` and `mass` for the Cartesian time `t` and proper time `tau` ($\tau$).
+
+### Familiar conventions
+
[…]
