### 2024-12-19 ce1ef28
finished the 'Vector objects' tutorial
