### 2024-12-19 035b79f
finished the 'SymPy expressions' tutorial
