+++ tests/test_notebooks.py
@@ -20,3 +20,23 @@ def common_kwargs(tmpdir):
 def test_object(common_kwargs):
     execution_dir = Path.cwd() / "docs" / "src"
     pm.execute_notebook(execution_dir / "object.ipynb", **common_kwargs)
+
+
+def test_numpy(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "numpy.ipynb", **common_kwargs)
+
+
+def test_awkward(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "awkward.ipynb", **common_kwargs)
+
+
+def test_numba(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "numba.ipynb", **common_kwargs)
+
+
+def test_sympy(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "sympy.ipynb", **common_kwargs)
