### 2024-12-19 67f7700
finished the 'NumPy arrays' tutorial
