### 2024-12-20 0764dd9
small fix in the contributor's guide
