### 2024-12-20 2a3daae
docs: rewrite Vector's documentation (#545)
* give the logo some margins

* remove the old; I'll use vector.readthedocs.io as a guide for refilling it

* switch from ReST to (Myst) Markdown

* style: pre-commit fixes

* skip test of non-existent notebook

* simplified the Overview page

* style: pre-commit fixes

* also mention geometric vs momentum

* style: pre-commit fixes

* added interfaces

* pre-commit

* rename 'talks' to 'presentations'

* finished the (public) reference documentation

* style: pre-commit fixes

* made stubs for tutorials

* use SVG and make README match readthedocs


[…]
