+++ src/vector/backends/numpy.py
@@ -1143,6 +1143,14 @@ class VectorNumpy2D(VectorNumpy, Planar, Vector2D, FloatArray):  # type: ignore[
         >>> vec
         VectorNumpy2D([(1.1, 2.1), (1.2, 2.2), (1.3, 2.3), (1.4, 2.4), (1.5, 2.5)],
                       dtype=[('x', '<f8'), ('y', '<f8')])
+
+        >>> import numpy as np
+        >>> import vector
+        >>> arr = np.array([(1.1, 2.1), (1.2, 2.2), (1.3, 2.3), (1.4, 2.4), (1.5, 2.5)],
+        ...               dtype=[('x', float), ('y', float)])
+        >>> arr.view(vector.VectorNumpy2D)
+        VectorNumpy2D([(1.1, 2.1), (1.2, 2.2), (1.3, 2.3), (1.4, 2.4), (1.5, 2.5)],
+                      dtype=[('x', '<f8'), ('y', '<f8')])
     """
 
     ObjectClass = vector.backends.object.VectorObject2D
@@ -1316,6 +1324,14 @@ class MomentumNumpy2D(PlanarMomentum, VectorNumpy2D):
         >>> vec
         MomentumNumpy2D([(1.1, 2.1), (1.2, 2.2), (1.3, 2.3), (1.4, 2.4), (1.5, 2.5)],
                         dtype=[('x', '<f8'), ('y', '<f8')])
+
+        >>> import numpy as np
+        >>> import vector
+        >>> arr = np.array([(1.1, 2.1), (1.2, 2.2), (1.3, 2.3), (1.4, 2.4), (1.5, 2.5)],
+        ...               dtype=[('px', float), ('py', float)])
+        >>> arr.view(vector.MomentumNumpy2D)
+        MomentumNumpy2D([(1.1, 2.1), (1.2, 2.2), (1.3, 2.3), (1.4, 2.4), (1.5, 2.5)],
+                        dtype=[('x', '<f8'), ('y', '<f8')])
     """
 
     ObjectClass = vector.backends.object.MomentumObject2D
@@ -1363,6 +1379,14 @@ class VectorNumpy3D(VectorNumpy, Spatial, Vector3D, FloatArray):  # type: ignore
         >>> vec = vector.VectorNumpy3D([(1.1, 2.1, 3.1), (1.2, 2.2, 3.2), (1.3, 2.3, 3.3), (1.4, 2.4, 3.4), (1.5, 2.5, 3.5)],
         ...               dtype=[('x', float), ('y', float), ('z', float)])
         >>> vec
+        VectorNumpy3D([(1.1, 2.1, 3.1), (1.2, 2.2, 3.2), (1.3, 2.3, 3.3), (1.4, 2.4, 3.4),
+                       (1.5, 2.5, 3.5)], dtype=[('x', '<f8'), ('y', '<f8'), ('z', '<f8')])
+
+        >>> import numpy as np
+        >>> import vector
+        >>> arr = np.array([(1.1, 2.1, 3.1), (1.2, 2.2, 3.2), (1.3, 2.3, 3.3), (1.4, 2.4, 3.4), (1.5, 2.5, 3.5)],
+        ...               dtype=[('x', float), ('y', float), ('z', float)])
+        >>> arr.view(vector.VectorNumpy3D)
         VectorNumpy3D([(1.1, 2.1, 3.1), (1.2, 2.2, 3.2), (1.3, 2.3, 3.3), (1.4, 2.4, 3.4),
[…]
