+++ tests/test_notebooks.py
@@ -17,6 +17,26 @@ def common_kwargs(tmpdir):
     }
 
 
-def test_intro(common_kwargs):
-    execution_dir = Path.cwd() / "docs" / "usage"
-    pm.execute_notebook(execution_dir / "intro.ipynb", **common_kwargs)
+def test_object(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "object.ipynb", **common_kwargs)
+
+
+def test_numpy(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "numpy.ipynb", **common_kwargs)
+
+
+def test_awkward(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "awkward.ipynb", **common_kwargs)
+
+
+def test_numba(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "numba.ipynb", **common_kwargs)
+
+
+def test_sympy(common_kwargs):
+    execution_dir = Path.cwd() / "docs" / "src"
+    pm.execute_notebook(execution_dir / "sympy.ipynb", **common_kwargs)
