+++ docs/api/compute/vector._compute.rst
@@ -1,18 +1,0 @@
-vector.\_compute package
-========================
-
-.. automodule:: vector._compute
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   lorentz/vector._compute.lorentz
-   planar/vector._compute.planar
-   spatial/vector._compute.spatial

[…]
