### 2024-12-22 5c317cb
style: pre-commit fixes
