### 2024-12-22 42afaa8
Rename `__azumthal` to `__azimuthal` (#548)
* typo fix

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
