### 2024-12-24 4b451e5
rm blacken-docs errors from _methods.py and docs/src
