+++ src/vector/_compute/spatial/cottheta.py
@@ -4,9 +4,10 @@
 # or https://github.com/scikit-hep/vector for details.
 
 """
+Property method `cottheta` for the cotangent theta coordinate of a spatial vector.
+
 .. code-block:: python
 
-    @property
     Spatial.cottheta(self)
 """
 
