+++ src/vector/backends/awkward.py
@@ -32,7 +32,9 @@ manages this non-strictness well.
 
 from __future__ import annotations
 
+import functools
 import numbers
+import operator
 import types
 import typing
 
@@ -966,6 +968,100 @@ class VectorAwkward:
 
         else:
             raise AssertionError(repr(returns))
+
+    def _wrap_dispatched_function(
+        self: AwkwardProtocol,
+        func: typing.Callable,  # type: ignore[type-arg]
+    ) -> typing.Callable:  # type: ignore[type-arg]
+        return awkward_transform(func)
+
+
+_placeholder = object()
+
+
+class bind(functools.partial):  # type: ignore[type-arg]
+    def __call__(self, *args: typing.Any, **keywords: typing.Any) -> typing.Any:
+        keywords = {**self.keywords, **keywords}
+        iargs = iter(args)
+        args = tuple(next(iargs) if arg is _placeholder else arg for arg in self.args)
+        return self.func(*args, *iargs, **keywords)
+
+
+class awkward_transform:
+    """Apply a batch of transformations in a single layout traversal"""
+
+    def __init__(self, func: typing.Callable) -> None:  # type: ignore[type-arg]
+        self.func = func
+
+    def __call__(self, *args: typing.Any, **kwargs: typing.Any) -> typing.Callable:  # type: ignore[type-arg]
+        # only pos args
+        assert not kwargs
+
+        # prepare the function and its args; we;re currently assuming non-nested input args
+        args2bind, awkward_arrays = [], []
+        n_orig_akarrays = 0
+        for arg in args:
+            if isinstance(arg, (ak.highlevel.Array, ak.highlevel.Record)):
+                n_orig_akarrays += 1
+                awkward_arrays.append(arg)
+                args2bind.append(_placeholder)
+            elif ak._regularize.is_array_like(arg):
+                awkward_arrays.append(ak.Array(ak.to_layout(arg)))
+                args2bind.append(_placeholder)
+            else:
+                args2bind.append(arg)
+
+        # this means we're working with awkward-arrays
+        if n_orig_akarrays > 0:
+
+            def transformer(
+                layouts: ak.contents.Content | tuple[ak.contents.Content, ...],
+                **kwargs: typing.Any,
+            ) -> ak.contents.Content:
+                if not isinstance(layouts, tuple):
+                    layouts = (layouts,)
+                if all(layout.is_numpy for layout in layouts):
+                    # setup parameter propagation, taken from awkward's broadcasting machinery
+                    options = kwargs["options"]
+                    rule = options.pop("broadcast_parameters_rule", None)
+                    try:
+                        parameters_factory = (
+                            ak._broadcasting.BROADCAST_RULE_TO_FACTORY_IMPL[rule]
+                        )
+                    except KeyError:
+                        raise ValueError(
+                            f"`broadcast_parameters_rule` should be one of {[str(x) for x in ak._broadcasting.BroadcastParameterRule]}, "
+                            f"but this routine received `{rule}`"
+                        ) from None
+
+                    # apply the function to the numpy arrays, first we need to 'partial it out' all non-awkward array arguments
+                    out_numpys = bind(self.func, *args2bind)(
+                        *(map(operator.attrgetter("data"), layouts))
+                    )
+                    # if the function returns a single array, wrap it in a tuple
+                    if not isinstance(out_numpys, tuple):
+                        out_numpys = (out_numpys,)
+                    # propagate parameters
+                    out_params = parameters_factory(
+                        (layout.parameters for layout in layouts), len(out_numpys)
+                    )
+                    # wrap the numpy arrays in awkward arrays
+                    out_arrays = tuple(
+                        ak.contents.NumpyArray(data=data, parameters=parameters)
+                        for data, parameters in zip(out_numpys, out_params)
+                    )
[…]
