+++ src/vector/_compute/spatial/eta.py
@@ -88,7 +88,9 @@ def dispatch(v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(v.lib, *v.azimuthal.elements, *v.longitudinal.elements),
+            v._wrap_dispatched_function(function)(
+                v.lib, *v.azimuthal.elements, *v.longitudinal.elements
+            ),
             returns,
             1,
         )
