### 2025-01-13 6377820
add a fast path for noops, otherwise ak.transform does a non-negligible overhead...
