+++ src/vector/_compute/spatial/eta.py
@@ -49,6 +49,9 @@ def xy_eta(lib, x, y, eta):
     return eta
 
 
+xy_eta.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
+
+
 def rhophi_z(lib, rho, phi, z):
     return lib.nan_to_num(
         lib.arcsinh(z / rho),
@@ -64,6 +67,9 @@ def rhophi_theta(lib, rho, phi, theta):
 
 def rhophi_eta(lib, rho, phi, eta):
     return eta
+
+
+rhophi_eta.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
 
 
 dispatch_map = {
