+++ src/vector/_compute/lorentz/tau.py
@@ -42,6 +42,9 @@ def xy_z_tau(lib, x, y, z, tau):
     return tau
 
 
+xy_z_tau.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
+
+
 def xy_theta_t(lib, x, y, theta, t):
     squared = tau2.xy_theta_t(lib, x, y, theta, t)
     return lib.copysign(lib.sqrt(lib.absolute(squared)), squared)
@@ -49,6 +52,9 @@ def xy_theta_t(lib, x, y, theta, t):
 
 def xy_theta_tau(lib, x, y, theta, tau):
     return tau
+
+
+xy_theta_tau.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
 
 
 def xy_eta_t(lib, x, y, eta, t):
@@ -60,6 +66,9 @@ def xy_eta_tau(lib, x, y, eta, tau):
     return tau
 
 
+xy_eta_tau.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
+
+
 def rhophi_z_t(lib, rho, phi, z, t):
     squared = tau2.rhophi_z_t(lib, rho, phi, z, t)
     return lib.copysign(lib.sqrt(lib.absolute(squared)), squared)
@@ -67,6 +76,9 @@ def rhophi_z_t(lib, rho, phi, z, t):
 
 def rhophi_z_tau(lib, rho, phi, z, tau):
     return tau
+
+
+rhophi_z_tau.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
 
 
 def rhophi_theta_t(lib, rho, phi, theta, t):
@@ -78,6 +90,9 @@ def rhophi_theta_tau(lib, rho, phi, theta, tau):
     return tau
 
 
+rhophi_theta_tau.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
+
+
 def rhophi_eta_t(lib, rho, phi, eta, t):
     squared = tau2.rhophi_eta_t(lib, rho, phi, eta, t)
     return lib.copysign(lib.sqrt(lib.absolute(squared)), squared)
@@ -85,6 +100,9 @@ def rhophi_eta_t(lib, rho, phi, eta, t):
 
 def rhophi_eta_tau(lib, rho, phi, eta, tau):
     return tau
+
+
+rhophi_eta_tau.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
 
 
 dispatch_map = {
