### 2025-01-15 78cd02f
perf: reduce python overhead for awkward backend (#554)
* awkward backend: reduce python overhead by applying multiple operations in one single broadcasting traversal

* add a fast path for noops, otherwise ak.transform does a non-negligible overhead...
