### 2025-01-21 9c6a55a
fix: lib for typetracers should use its nplike when possible
