+++ tests/backends/test_awkward.py
@@ -5,6 +5,7 @@
 
 from __future__ import annotations
 
+import functools
 import importlib.metadata
 import numbers
 
@@ -26,97 +27,120 @@ awkward_without_record_reducers = packaging.version.Version(
 ) < packaging.version.Version("2.2.3")
 
 
-def test_dimension_conversion():
+def _assert(bool1, bool2, backend):
+    # for 'typetracer' backend we pass through (no-op)
+    # for 'cpu' we want to make sure the bool evaluates to True
+    if backend == "cpu":
+        if isinstance(bool1, ak.Array):
+            bool1 = bool1.tolist()
+        if isinstance(bool2, ak.Array):
+            bool2 = bool2.tolist()
+        else:
+            assert bool1 == bool2
+
+
+@pytest.mark.parametrize("backend", ["cpu", "typetracer"])
+def test_dimension_conversion(backend):
+    assert_backend = functools.partial(_assert, backend=backend)
+
     # 2D -> 3D
     vec = vector.Array(
         [
             [{"x": 1, "y": 1.1}, {"x": 2, "y": 2.1}],
             [],
-        ]
+        ],
+        backend=backend,
     )
-    # test alias
-    assert ak.all(vec.to_3D(z=1).z == 1)
-    assert ak.all(vec.to_3D(eta=1).eta == 1)
-    assert ak.all(vec.to_3D(theta=1).theta == 1)
 
-    assert ak.all(vec.to_Vector3D(z=1).x == vec.x)
-    assert ak.all(vec.to_Vector3D(z=1).y == vec.y)
+    # test alias
+    assert_backend(ak.all(vec.to_3D(z=1).z == 1), True)
+    assert_backend(ak.all(vec.to_3D(eta=1).eta == 1), True)
+    assert_backend(ak.all(vec.to_3D(theta=1).theta == 1), True)
+
+    assert_backend(vec.to_Vector3D(z=1).x, vec.x)
+    assert_backend(vec.to_Vector3D(z=1).y, vec.y)
 
     # 2D -> 4D
-    assert ak.all(vec.to_Vector4D(z=1, t=1).t == 1)
-    assert ak.all(vec.to_Vector4D(z=1, t=1).z == 1)
-    assert ak.all(vec.to_Vector4D(eta=1, t=1).eta == 1)
-    assert ak.all(vec.to_Vector4D(eta=1, t=1).t == 1)
-    assert ak.all(vec.to_Vector4D(theta=1, t=1).theta == 1)
-    assert ak.all(vec.to_Vector4D(theta=1, t=1).t == 1)
-    assert ak.all(vec.to_Vector4D(z=1, tau=1).z == 1)
-    assert ak.all(vec.to_Vector4D(z=1, tau=1).tau == 1)
-    assert ak.all(vec.to_Vector4D(eta=1, tau=1).eta == 1)
-    assert ak.all(vec.to_Vector4D(eta=1, tau=1).tau == 1)
-    assert ak.all(vec.to_Vector4D(theta=1, tau=1).theta == 1)
-    assert ak.all(vec.to_Vector4D(theta=1, tau=1).tau == 1)
+    assert_backend(ak.all(vec.to_Vector4D(z=1, t=1).t == 1), True)
+    assert_backend(ak.all(vec.to_Vector4D(z=1, t=1).z == 1), True)
+    assert_backend(ak.all(vec.to_Vector4D(eta=1, t=1).eta == 1), True)
+    assert_backend(ak.all(vec.to_Vector4D(eta=1, t=1).t == 1), True)
+    assert_backend(ak.all(vec.to_Vector4D(theta=1, t=1).theta == 1), True)
+    assert_backend(ak.all(vec.to_Vector4D(theta=1, t=1).t == 1), True)
[…]
