+++ src/vector/backends/awkward.py
@@ -649,7 +649,11 @@ class VectorAwkward:
 
     @property
     def lib(self):  # type:ignore[no-untyped-def]
-        return _lib(module=numpy, nplike=self.layout.backend.nplike)  # type:ignore[attr-defined]
+        if (
+            nplike := self.layout.backend.nplike  # type:ignore[attr-defined]
+        ) is ak._nplikes.typetracer.TypeTracer.instance():
+            return _lib(module=numpy, nplike=nplike)
+        return numpy
 
     def _wrap_result(
         self: AwkwardProtocol,
