### 2025-01-22 24fd9f2
improve comment for assertion helper function
