### 2025-01-22 96e71b3
fix: lib for typetracers (#558)
* fix: lib for typetracers should use its nplike when possible

* restrict fix to typetracer backend

* improve comment for assertion helper function
