+++ .pre-commit-config.yaml
@@ -40,6 +40,7 @@ repos:
     rev: v2.4.0
     hooks:
       - id: codespell
+        args: ["-LHEP"]
         exclude: ^docs/usage/intro.ipynb$
 
   - repo: https://github.com/rbubley/mirrors-prettier
