### 2025-02-17 bd1d438
chore: update pre-commit hooks
updates:
- [github.com/rbubley/mirrors-prettier: v3.5.0 → v3.5.1](https://github.com/rbubley/mirrors-prettier/compare/v3.5.0...v3.5.1)
