### 2025-02-18 #7791 @editorialbot
[REVIEW]: Vector: JIT-compilable mathematical manipulations of ragged Lorentz vectors
**Submitting author:** @Saransh-cpp (<a href="http://orcid.org/0000-0003-3046-7675">Saransh Chopra</a>)
**Repository:** https://github.com/scikit-hep/vector
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.6.2
