### 2025-02-20 c6f675f
Fix citation, better explanations, and fix English
