### 2025-02-20 9486e98
Fix citation, better explanations, and fix English
