+++ paper/paper.bib
@@ -221,3 +221,193 @@ title = {{PyTorch 2: Faster Machine Learning Through Dynamic Python Bytecode Tra
 url = {https://pytorch.org/assets/pytorch2-2.pdf},
 year = {2024}
 }
+
+@software{hepvector,
+  author = {Henry Schreiner},
+  title = {{HEPvector}: NumPy based vectors for general purpose calculation and physics},
+  url = {https://github.com/henryiii/hepvector},
+}
+
+@software{LorentzVectorHEP:2023,
+  author = {Jerry Ling and Moritz Terwissen and Graeme A Stewart},
+  title = {LorentzVectorHEP.jl: x, y, z, t and pt, eta, phi, M Lorentz Vectors as used in collider experiments at CERN and other places},
+  url = {https://github.com/JuliaHEP/LorentzVectorHEP.jl},
+  version = {0.1.16},
+  year = {2023},
+}
+
+@article{Held:2022sfw,
+    author = "Held, Alexander and Shadura, Oksana",
+    title = "{The IRIS-HEP Analysis Grand Challenge}",
+    doi = "10.22323/1.414.0235",
+    journal = "PoS",
+    volume = "ICHEP2022",
+    pages = "235",
+    month = "11",
+    year = "2022"
+}
+
+@book{Aberle:2749422,
+      author        = "Aberle, O. and Béjar Alonso, I and Brüning, O and
+                       Fessia, P and Rossi, L and Tavian, L and Zerlauth, M and
+                       Adorisio, C. and Adraktas, A. and Ady, M. and Albertone, J.
+                       and Alberty, L. and Alcaide Leon, M. and Alekou, A. and
+                       Alesini, D. and Ferreira, B. Almeida and Lopez, P. Alvarez
+                       and Ambrosio, G. and Andreu Munoz, P. and Anerella, M. and
+                       Angal-Kalinin, D. and Antoniou, F. and Apollinari, G. and
+                       Apollonio, A. and Appleby, R. and Arduini, G. and Alonso,
+                       B. Arias and Artoos, K. and Atieh, S. and Auchmann, B. and
+                       Badin, V. and Baer, T. and Baffari, D. and Baglin, V. and
+                       Bajko, M. and Ball, A. and Ballarino, A. and Bally, S. and
+                       Bampton, T. and Banfi, D. and Barlow, R. and Barnes, M. and
+                       Barranco, J. and Barthelemy, L. and Bartmann, W. and
+                       Bartosik, H. and Barzi, E. and Battistin, M. and
+                       Baudrenghien, P. and Alonso, I. Bejar and Belomestnykh, S.
+                       and Benoit, A. and Ben-Zvi, I. and Bertarelli, A. and
+                       Bertolasi, S. and Bertone, C. and Bertran, B. and Bestmann,
+                       P. and Biancacci, N. and Bignami, A. and Bliss, N. and
+                       Boccard, C. and Body, Y. and Borburgh, J. and Bordini, B.
+                       and Borralho, F. and Bossert, R. and Bottura, L. and
+                       Boucherie, A. and Bozzi, R. and Bracco, C. and Bravin, E.
+                       and Bregliozzi, G. and Brett, D. and Broche, A. and
+                       Brodzinski, K. and Broggi, F. and Bruce, R. and Brugger, M.
+                       and Brüning, O. and Buffat, X. and Burkhardt, H. and
+                       Burnet, J. and Burov, A. and Burt, G. and Cabezas, R. and
+                       Cai, Y. and Calaga, R. and Calatroni, S. and Capatina, O.
+                       and Capelli, T. and Cardon, P. and Carlier, E. and Carra,
[…]
