+++ paper/paper.md
@@ -64,9 +64,9 @@ high energy physics experiments.
 
 Vector is one of the few Lorentz vector libraries that offer
 a Pythonic interface but a compiled computational backend, with the others
-being Coffea's vector module (depends on vector), PyROOT's [@root:2020]
+being Coffea's vector module (depends on Vector), PyROOT's [@root:2020]
 LorentzVectors and TLorentzVector classes, and HEPvector [@hepvector] (deprecated
-in favor of vector). Although vector was written with high energy physics in mind,
+in favor of Vector). Although Vector was written with high energy physics in mind,
 it is a general-purpose library that can be used for any scientific or engineering
 application. The library houses a set of diverse backends, three numerical backends
 for experimental physicists and one symbolic backend for theoretical physicists.
@@ -77,11 +77,11 @@ These backends are:
 - a SymPy [@Meurer:2017] backend for symbolic computations, and
 - an Awkward [@Pivarski:2018] backend for computations on ragged collection-type data
 
-Moreover, vector is the first Lorentz vector library to offer multiple
+Moreover, Vector is the first Lorentz vector library to offer multiple
 computational backends, as well as both numerical and symbolic backends. Furthermore,
-akin to PyROOT and LorentzVectorHEP.jl [@LorentzVectorHEP:2023], vector supports
+akin to PyROOT and LorentzVectorHEP.jl [@LorentzVectorHEP:2023], Vector supports
 just-in-time compilation through Numba extensions [@lam:2015], implemented for both
-the Object and Awkward backends. Vector also includes support for JAX [@Bradbury:2018]
+the object and Awkward backends. Vector also includes support for JAX [@Bradbury:2018]
 and Dask [@rocklin:2015] for the Awkward backend, enabling the library to support
 automatic differentiation and parallel computing, which are required for introducing
 automatic differentiation in Analysis Grand Challenge [@Held:2022sfw] and to meet
@@ -89,7 +89,7 @@ the computational needs of High Luminosity LHC [@Aberle:2749422].
 
 ## Impact
 
-Besides PyROOT's LorentzVectors and TLorentzVector, vector has
+Besides PyROOT's LorentzVectors and TLorentzVector, Vector has
 become a popular choice for mathematical manipulations in Python based high energy
 physics analysis pipelines. Along with being utilized directly in
 analysis pipelines at LHC [@Kling:2023; @Held:2024; @Qu:2022], the library is also
@@ -102,8 +102,8 @@ often used in non high energy physics use cases.
 
 # Acknowledgements
 
-The work on vector was supported by NSF cooperative agreements OAC-1836650
+The work on Vector was supported by NSF cooperative agreements OAC-1836650
 (IRIS-HEP) and PHY-2323298 (IRIS-HEP). We would additionally like to thank the
-contributors of vector and the Scikit-HEP community for their support.
+contributors of Vector and the Scikit-HEP community for their support.
 
 # Reference
