### 2025-03-02 #7791 @Saransh-cpp
Thanks for the detailed review, @rafaelab!

> - The documentation should necessarily specify things like metric conventions for four-vectors (that also changes definitions of time-like and space-like). People working in general relativity tend to use the other convention (and some in HEP as well).
> - It is also important to explicitly add to the documentation if boosts are active or passive.

I am hoping @henryiii or @jpivarski can help me with these (I hope it is okay to ping you both on this review).

1. If I am not wrong, Vector follows the metric conventions is `(-,-,-,+)`. @rafaelab could you please clarify if by changing the definitions of time-like and space-like you mean updating the docstrings of `is_timelike` and `is_spacelike` to highlight this convention?
2. Given that we follow ROOT, the boosts should be active, as formalised in - https://root.cern/topical/GenVector.pdf. However, the PDF also says that the conventions for Euler angles is passive. Should we just link to the PDF to reference the active/passive conventions Vector follows?

> - I'd be interested in seeing two performance benchmarks comparing to ROOT's TLorentzVector, a simple one, and one with Numba.

There was an attempt for writing benchmarks against ROOT, but the efforts stalled a few years back - https://github.com/scikit-hep/vector/pull/36. I will ping the other maintainers and check if someone is interested in finishing that PR. We already have a Numba notebook, and I have created a PR to showcase simple comparisons between JIT and non-JIT calls - https://github.com/scikit-hep/vector/pull/565.

> - Allow the user to choose the metric convention (for people not in HEP, this is very useful)
> - A constructor such as vector.obj(rho=1) that assumes theta=0 and phi=0 automatically would be very handy.
> - Along these lines, a constructor like VectorObject3D(x=1) automatically setting y and z to 0 would be useful.

It would be amazing if you can open standalone issues for these bullet points in the vector repository. I (or someone else) will take a look at them as soon as possible!
