+++ tests/root/test_vector.py
@@ -1,5 +1,6 @@
 # This test code was written by the `hypothesis.extra.ghostwriter` module
 # and is provided under the Creative Commons Zero public domain dedication.
+from __future__ import annotations
 
 import pytest
 
