### 2025-03-09 #7791 @rafaelab
Thanks for the reply, @Saransh-cpp

> If I am not wrong, Vector follows the metric conventions is (-,-,-,+). @rafaelab could you please clarify if by changing the definitions of time-like and space-like you mean updating the docstrings of is_timelike and is_spacelike to highlight this convention?

I checked and the code does follow the "mostly-minus" metric convention. My suggestion is just to make this explicit in the documentation. The mostly-plus convention is also widely used, and if users (especially students) use the code without paying attention to this, it could be problematic.
I don't think there is a need to update the docstrings of `is_timelike` and `is_spacelike`, but these functions do depend on the choice of metric convention.

> There was an attempt for writing benchmarks against ROOT, but the efforts stalled a few years back - https://github.com/scikit-hep/vector/pull/36. I will ping the other maintainers and check if someone is interested in finishing that PR. We already have a Numba notebook, and I have created a PR to showcase simple comparisons between JIT and non-JIT calls - https://github.com/scikit-hep/vector/pull/565.

The JIT/non-JIT comparison is great, thanks!
There is actually no need to compare with ROOT, it is just out of curiosity.

In summary, the only thing I think is need is to add these notes to the documentation.
