### 2025-03-09 a5cfcc2
docs: add metric and transformation conventions
