### 2025-03-11 8254275
add more physics details
Co-authored-by: Ianna Osborne <ianna.osborne@cern.ch>
