### 2025-03-14 5cb2237
docs: add metric and transformation conventions (#568)
* docs: add metric and transformation conventions

* be more explicit about coordinates' order

* add more physics details

Co-authored-by: Ianna Osborne <ianna.osborne@cern.ch>

---------

Co-authored-by: Ianna Osborne <ianna.osborne@cern.ch>
