### 2025-04-07 d5e954a
chore: update pre-commit hooks (#574)
