### 2025-04-21 f650f5d
chore: update pre-commit hooks (#577)
