### 2025-04-22 #7791 @adonath
@Saransh-cpp You might proceed with making a new release of the software (incl. the latest changes from the review process) and upload the version to an archive such as Zenodo. Please also check the other ToDo items in https://github.com/openjournals/joss-reviews/issues/7791#issuecomment-2821510208

Thanks!
