### 2025-04-22 #7791 @Saransh-cpp
Thanks, @adonath! I have created a new release and archive on Zenodo. Should the author list be the same on Zenodo and the paper or should the citation be updated once the paper is accepted?
