### 2025-04-22 #7791 @adonath
>  Should the author list be the same on Zenodo and the paper

Yes, both should be equivalent. The author list on Zenodo seems to miss a few co-authors. Can you please adapt and re-release?
