### 2025-04-22 #7791 @danielskatz
FYI, this just requires a change in the metadata of the existing deposit. It does not require a new deposit or a new DOI
