### 2025-04-30 0a9db8a
fixing sympy tests
