+++ tests/compute/sympy/lorentz/test_tau2.py
@@ -74,9 +74,10 @@ def test_xy_eta_t():
         vector.backends.sympy.TemporalSympyT(t),
     )
     # TODO: the expression blows up on simplifying?
-    assert vec.tau2 == t**2 - 0.25 * (
+    expected = t**2 - 0.25 * (
         1 + sympy.exp(-2 * sympy.asinh(z / sympy.sqrt(x**2 + y**2)))
     ) ** 2 * (x**2 + y**2) * sympy.exp(2 * sympy.asinh(z / sympy.sqrt(x**2 + y**2)))
+    assert sympy.simplify(expected - vec.tau2) == 0
     assert vec.tau2.subs(values).evalf() == pytest.approx(275)
 
 
