+++ tests/compute/sympy/lorentz/test_t2.py
@@ -92,11 +92,12 @@ def test_xy_eta_tau():
         ),
     )
     # TODO: the expression blows up on simplifying?
-    assert vec.t2 == 0.25 * (
+    expected = 0.25 * (
         1 + sympy.exp(-2 * sympy.asinh(z / sympy.sqrt(x**2 + y**2)))
     ) ** 2 * (x**2 + y**2) * sympy.exp(
         2 * sympy.asinh(z / sympy.sqrt(x**2 + y**2))
     ) + sympy.Abs(-(t**2) + x**2 + y**2 + z**2)
+    assert sympy.simplify(vec.t2 - expected) == 0
     assert vec.t2.subs(values).evalf() == pytest.approx(400)
 
 
@@ -167,7 +168,8 @@ def test_rhophi_eta_tau():
         ),
     )
     # TODO: the expression blows up on simplifying?
-    assert vec.t2 == 0.25 * rho**2 * (
+    expected = 0.25 * rho**2 * (
         1 + sympy.exp(-2 * sympy.asinh(z / rho))
     ) ** 2 * sympy.exp(2 * sympy.asinh(z / rho)) + sympy.Abs(rho**2 - t**2 + z**2)
+    assert sympy.simplify(vec.t2 - expected) == 0
     assert vec.t2.subs(values).evalf() == pytest.approx(400)
