+++ tests/compute/sympy/lorentz/test_t.py
@@ -92,13 +92,14 @@ def test_xy_eta_tau():
         ),
     )
     # TODO: the expression blows up on simplifying?
-    assert vec.t == sympy.sqrt(
+    expected = sympy.sqrt(
         0.25
         * (1 + sympy.exp(-2 * sympy.asinh(z / sympy.sqrt(x**2 + y**2)))) ** 2
         * (x**2 + y**2)
         * sympy.exp(2 * sympy.asinh(z / sympy.sqrt(x**2 + y**2)))
         + sympy.Abs(-(t**2) + x**2 + y**2 + z**2)
     )
+    assert sympy.simplify(vec.t - expected) == 0
     assert vec.t.subs(values).evalf() == pytest.approx(20)
 
 
@@ -173,11 +174,12 @@ def test_rhophi_eta_tau():
         ),
     )
     # TODO: the expression blows up on simplifying?
-    assert vec.t == sympy.sqrt(
+    expected = sympy.sqrt(
         0.25
         * rho**2
         * (1 + sympy.exp(-2 * sympy.asinh(z / rho))) ** 2
         * sympy.exp(2 * sympy.asinh(z / rho))
         + sympy.Abs(rho**2 - t**2 + z**2)
     )
+    assert sympy.simplify(expected - vec.t) == 0
     assert vec.t.subs(values).evalf() == pytest.approx(20)
