### 2025-05-12 #7791 @eduardo-rodrigues
Dear colleagues. It's not quite in the spirit of JOSS to have a review lasting close to 3 months, I hope. This is more than it takes for a typical peer reviewed paper in physics ;-). Can you let us know what is preventing us from concluding? I appreciate that all efforts are done in "free time", no worries. Thank you and best wishes.
