### 2025-05-23 #7791 @kyleniemeyer
@Saransh-cpp I made a few edits to the paper in https://github.com/scikit-hep/vector/pull/589, could you review and merge? I will then accept and publish your submission.
