### 2025-05-23 d96985a
Updates JOSS paper (#589)
* Updates JOSS paper

* Updates JOSS bibliography
