+++ paper/paper.bib
@@ -18,7 +18,6 @@
  pages         = {357--362},
  doi           = {10.1038/s41586-020-2649-2},
  publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
 @software{Gray:2023,
@@ -50,15 +49,17 @@ year = {2018}
 }
 
 @inproceedings{lam:2015,
-  title={Numba: A llvm-based python jit compiler},
+  title={Numba: A {LLVM}-based {P}ython {JIT} compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  number={7},
+  doi={10.1145/2833157.2833162}
 }
 
 @article{Meurer:2017,
-    title = {SymPy: symbolic computing in Python},
+    title = {{SymPy}: symbolic computing in {P}ython},
     author = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k, Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K. and Singh, Sartaj and Rathnayake, Thilina and Vig, Sean and Granger, Brian E. and Muller, Richard P. and Bonazzi, Francesco and Gupta, Harsh and Vats, Shivam and Johansson, Fredrik and Pedregosa, Fabian and Curry, Matthew J. and Terrel, Andy R. and Rou\v{c}ka, \v{S}t\v{e}p\'{a}n and Saboo, Ashutosh and Fernando, Isuru and Kulal, Sumith and Cimrman, Robert and Scopatz, Anthony},
     year = 2017,
     month = jan,
@@ -75,10 +76,9 @@ year = {2018}
 }
 
 @article{Kling:2023,
-   title={FLArE up dark sectors with EM form factors at the LHC forward physics facility},
+   title={{FLArE} up dark sectors with {EM} form factors at the {LHC} forward physics facility},
    volume={987},
    ISSN={0550-3213},
-   url={http://dx.doi.org/10.1016/j.nuclphysb.2023.116103},
    DOI={10.1016/j.nuclphysb.2023.116103},
    journal={Nuclear Physics B},
    publisher={Elsevier BV},
@@ -88,12 +88,12 @@ year = {2018}
 
 @article{Held:2024,
     author = "Held, Alexander and Kauffman, Elliott and Shadura, Oksana and Wightman, Andrew",
-    title = "{Physics analysis for the HL-LHC: Concepts and pipelines in practice with the Analysis Grand Challenge}",
+    title = {Physics analysis for the {HL}-{LHC}: Concepts and pipelines in practice with the {A}nalysis {G}rand {C}hallenge},
     eprint = "2401.02766",
     archivePrefix = "arXiv",
     primaryClass = "hep-ex",
     doi = "10.1051/epjconf/202429506016",
-    journal = "EPJ Web Conf.",
+    journal = "EPJ Web of Conferences",
     volume = "295",
     pages = "06016",
     year = "2024"
@@ -112,13 +112,13 @@ year = {2018}
 
 @article{Brehmer:2020,
       author         = "Brehmer, Johann and Kling, Felix and Espejo, Irina and Cranmer, Kyle",
-      title          = "{MadMiner: Machine learning-based inference for particle physics}",
-      journal        = "Comput. Softw. Big Sci.",
+      title          = {{MadMiner}: Machine learning-based inference for particle physics},
+      journal        = {Computing and Software for Big Science},
[…]
