### 2025-05-23 #7791 @Saransh-cpp
Thanks, @kyleniemeyer, merged!
