+++ paper/paper.bib
@@ -217,7 +217,7 @@ booktitle = {29th ACM International Conference on Architectural Support for Prog
 doi = {10.1145/3620665.3640366},
 month = apr,
 publisher = {ACM},
-title = {PyTorch} 2: {F}aster Machine Learning Through Dynamic {P}ython Bytecode Transformation and Graph Compilation},
+title = {{PyTorch} 2: {F}aster Machine Learning Through Dynamic {P}ython Bytecode Transformation and Graph Compilation},
 url = {https://pytorch.org/assets/pytorch2-2.pdf},
 year = {2024}
 }
